requirements.txt
django
djangorestframework
pillow
twilio

